rename: features 폴더로 재구성
